refactor parquet reader
