+++ be/src/exec/parquet_builder.h
@@ -21,49 +21,17 @@
 
 #pragma once
 
-#include <arrow/api.h>
-#include <arrow/buffer.h>
-#include <arrow/io/api.h>
-#include <arrow/io/file.h>
-#include <arrow/io/interfaces.h>
-#include <parquet/api/reader.h>
-#include <parquet/api/writer.h>
-#include <parquet/arrow/reader.h>
-#include <parquet/arrow/writer.h>
-#include <parquet/exception.h>
-
 #include <cstdint>
 #include <map>
 #include <string>
 
 #include "common/status.h"
 #include "exec/file_builder.h"
-#include "gen_cpp/FileBrokerService_types.h"
-#include "gen_cpp/PlanNodes_types.h"
-#include "gen_cpp/Types_types.h"
 
 namespace starrocks {
 
 class ExprContext;
-class FileWriter;
-
-class ParquetOutputStream : public arrow::io::OutputStream {
-public:
-    ParquetOutputStream(WritableFile* _writable_file);
-    ~ParquetOutputStream() override;
-
-    arrow::Status Write(const void* data, int64_t nbytes) override;
-    // return the current write position of the stream
-    arrow::Result<int64_t> Tell() const override;
-    arrow::Status Close() override;
-
-    bool closed() const override { return _is_closed; }
-
-private:
-    WritableFile* _writable_file; // not owned
-    int64_t _cur_pos = 0;         // current write position
-    bool _is_closed = false;
-};
+class ParquetOutputStream;
 
 // a wrapper of parquet output stream
 class ParquetBuilder : public FileBuilder {
