+++ be/src/exec/vectorized/parquet_scanner.cpp
@@ -7,7 +7,6 @@
 
 #include "column/column_helper.h"
 #include "env/env_util.h"
-#include "exec/parquet_reader.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
 #include "runtime/stream_load/load_stream_mgr.h"
